requirements.txt
altgraph==0.17.4
azure-core==1.32.0
azure-identity==1.21.0
azure-keyvault-secrets==4.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.2
httplib2==0.22.0
idna==3.10
isodate==0.7.2
msal==1.32.0
msal-extensions==1.2.0
numpy==2.2.4
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
pefile==2023.2.7
portalocker==2.10.1
proto-plus==1.26.1
protobuf==6.30.2
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
PyJWT==2.10.1
pyodbc==5.2.0
pyparsing==3.2.3
PySide6==6.9.0
PySide6_Addons==6.9.0
PySide6_Essentials==6.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pywin32==309
pywin32-ctypes==0.2.3
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
schedule==1.2.2
setuptools==76.0.0
shiboken6==6.9.0
six==1.17.0
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
